+++ producers/ni/include/NiController.hh
@@ -79,7 +79,7 @@ private:
 	char Buffer_data[7000];
 	char Buffer_length[7000];
 
-	unsigned long data_trans_addres;// = INADDR_NONE;
+	uint32_t data_trans_addres;// = INADDR_NONE;
 	SOCKET sock_config;
 	SOCKET sock_datatransport;
 	int numbytes;
